third_party/blink/web_tests/platform/generic/fast/css/getPropertyValue-webkit-text-stroke-expected.txt

Bug 111018: getPropertyValue for -webkit-text-stroke returns null, should compute the shorthand value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS webkitTextStrokeValue("text-stroke1") is "thin red"
PASS webkitTextStrokeValue("text-stroke2") is "green"
PASS webkitTextStrokeValue("text-stroke3") is "thin"
PASS webkitTextStrokeValue("text-stroke4") is "1px red"
PASS webkitTextStrokeValue("text-stroke5") is "1px black"
NOTE: 'foo' is an illegal CSS value for '-webkit-text-stroke'.
PASS webkitTextStrokeValue("text-stroke6") is ""
NOTE: If only few longhand properties are specified, getPropertyValue for shorthand property returns empty string.
PASS webkitTextStrokeValue("text-stroke7") is ""
PASS successfullyParsed is true

TEST COMPLETE
